requirements.txt
pytorch
matplotlib
numpy
pandas
tqdm